pyproject.toml
[tool.poetry]
name = "cookai"
version = "0.1.0"
description = ""
authors = ["2017250297-choi <97ckdtn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]

python = ">=3.8,<3.9.dev0"
django = "^4.2.2"
django-rest-framework = "^0.1.0"
django-dotenv = "^1.4.2"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^4.0.0"
requests = "^2.31.0"
opencv-python = "^4.7.0.72"
django-crontab = "^0.7.1"
scikit-learn = "^1.2.2"
scipy = "^1.10.1"
django-taggit = "^4.0.0"
django-taggit-serializer = "^0.1.7"
ratelimiter = "^1.2.0.post0"
pandas="^2.0.2"


[tool.poetry.group.dev.dependencies]
setuptools = "^67.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"